+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 - Add test for ArnaApp
 - Add test for ArnaScaffold
 - Move animations out of Arna
+- [BREAKING] Replace ArnaBottomBar with ArnaNavigationBar
+- [BREAKING] Replace ArnaBottomBarItem with ArnaNavigationBarItem
+- Add test for ArnaNavigationBar
 
 ## 1.0.6